Upgrade container release version
Since I'm a magit junky, and this is the easiest way to get
a version of emacs that's recent enough to run that.

Not a good reason to move from the LTS on anything that would be
deployed in the real world, but this is really my template for building
development environments anyway. So run with it for now.

At this rate, we'll be into the next LTS before it's in any shape for
anyone else to actually look at anyway.