Minor tweaks, working through manually
Main change seems to be that the latest version of ubuntu changes the
name of gpg2 back to gpg.

Added some ssh config rules to make the connections more
secure and happen faster (hopefully).

Added another frereth repo to the mix, since I obviously didn't have
enough to start with.