Get basic web worker working
More or less. I'm getting messages spewing between my worker and the
card that's triggering its work. It's actually bad enough that it's
crashing Chrome.

I'm pretty sure I see the problem, but enough work/research went into
getting this far that I wanted to be sure to milestone it before I risk
mangling something that's so close to working.