Start converting to a bash script
Well, trying. I'm mixing layers about where some of the most important
commands are run. But it's easier than trying to remember how ansible
actually works.

It isn't something I'd try to use for pushing into production, and it's
still just in the baby-step phases.

But it is making forward progress.