Wrap up pseudo-bash script for building container
It isn't exactly runnable, but at least it's documentation. Also
tweaked some details in the ansible play books, though this
approach doesn't exactly seem viable.

Then again, it's perfect for setting up a VM instead of an LXC. I
should probably be planning on that as the distribution mechanism
anyway.