Rough sketch of an App outline (WIP)
At least, that's what I think this is. The skeleton/template for howto
write your own app.

Of course, I'm making it up as I go. And I'm halfway guessing. This is
a half-baked idea that I must have started noodling on at one point and
then forgotten.

At this point it's just something I ran across as I was doing project
archaeology for backup purposes.