Pull in the rest of Deraen/saapas (WIP)
I think all the pieces are in place now, though not tested.

I had to translate a few things on the server side, because of library
changes.

For now, I'm skipping the CSS preprocessors because that seems like
something that would be more appropriate for the individual apps.

That part's going to be interesting.

Added a basic README.