Start debugging the REPL from Docker
This was mostly about hard-coding the different ports involved and
then setting up a run script that exposes them.

This version is annoying, because a lot of libraries still get
downloaded in when you run the container instead of at build time.

And trying to run cider-connect-sibling-cljs fails because it insists
on listening to localhost rather than 0.0.0.0.