Add basic log generation/consumption
This mocks up a server piece that sends log messages to tap>

And has a skeleton implementation for a client piece that consumes
those logs.

It's really just a first step toward sending those values to the
frontend log-viewer via a web socket so its worker can format them for
the renderer.