Start generating crypto-key for forked PIDs (WIP)
On the browser side, each "process" needs a unique identifier for the
sake of message routing.

It needs to be crypto-secure because we can't trust any of the puzzle
pieces.

We could use "standard" UUID4s for this, but those need crypto-strong
randomness for these purposes anyway, so we might as just skip
straight to the chase and handle it this way.

Also, started refactoring the actual connection/login steps into
something similar to what happens during a linux shell login. This
part is almost all comments.

But the pieces do live in slightly smaller functions now.

Started planning to use transit to shuffle data back and forth. At
least theoretically, it should be faster than EDN.

This version fails because the Renderer/Session key that arrives
doesn't quite match what's expected, even though all the numbers seem
to look right.

This particular problem is likely a type mismatch.