Find/fix websocket bugs
I needed to convert the ByteArrayOutputStream to a byte array before
sending it. That was killing the web socket with a secretive 1006
error code.

Once the browser side received that message, it failed to decode it
because transit operates on strings rather than byte arrays (and blobs
weren't even a little bit viable).