Expand fork-shell! to extract public key as JWK
Then extend that to use the map as a PID that's a key to both the
WebWorker and the full key pair.

Break up the chain of .then calls into standalone functions to try to
make it a bit more readable.

Also, did a bit of rename refactoring and cleaned up a little of the
debugging cruft.