Notify webserver before browser forks shell (WIP)
There's a lot going on here.

A bit of it was just puttering around with minor refactorings, trying
to decide what and how.

Now the browser is going to fail to send the webserver a notification
that it's ready to fork the shell. This fails because I'm not
serializing the message yet.

On the web server side, I've set up pieces that seem like they ought
to be able to handle that message. More importantly, I think the basic
idea should work out OK for handling pretty much any message along
these lines.

This is dangerous territory, and this doesn't even really qualify as a
first draft. It's totally untested.

But it compiles. And there are enough changes here that I don't want
to make this commit any bigger than I must.