Pull client.propagate out of main system
That's something that should start when a World on the browser starts
listening for those log messages.

For that matter, it should also stop sending when the last World stops
listening, but I haven't touched this in the code yet.

Switch the webserver dispatch key from ::fork to ::forking. Add
another handler for ::forked. These are important lifecycle stages
that need to be coordinated (for things like starting/stopping that
log propagation).

On the browser, refactored send- and recv- message! into their own
lexical closures with lamport clocks and transit
readers/writers. Mostly because they seem like pieces that are begging
to be copy/pasted. But also because those are abstractions that are
worth isolating.

Added the idea of sending the command you're about to fork along with
the notification that a fork is about to happen. I'm not doing
anything with that yet, but it seems like a good idea.

Fixed a misnamed spec in browser's fork-login! function (which still
doesn't actually do anything).