Get ::forking message from browser to handler
This was pretty much all debug print statements. Mostly because I
thought the problem was probably something to do with the client
send.

Added a default dispatch! method to help identify the problem.

That was an extra layer in the response wrapper tree. The part I
thought was the body is underneath the :frereth/body key.

Knocked out a TODO item to rename the wrap function to serialize.

Added some notes about how the browser really should handle worker
spawning to try to protect against rogue client DoS attacks.

Did fix a typo on the browser: had send-message! misspelled as
send-mesage!

Refactored the promise handler for exporting the worker's public key
(and then building the worker) into its own function.