Bandaid a circular dependency error
Between client.propagate and renderer.lib. propagate definitely needs
to be able to call a function that belongs in lib. But lib describes a
system that uses a function in propagate.

For now, I don't have to worry about this, because I haven't gotten to
the point where I actually need to create that system. But I will
soon, so this won't help for long.

Did some research into SRP. Added dependencies on a server-side
library for that, JWT (and especially JWK) handling, and also for
cleaner promise handling.

Also added a license notice to the README.

Renamed renderer's post-message to post-message! Added a declare to
the top of the file so internal implementatations can use it.