Debug the forking-world/ACK handshake (WIP)
There are also a bunch of whitespace changes from a recommended lint
fix, and a couple of renaming refactors, so it's easy to miss the
pieces that actually matter.

renderer/lib.clj line 170: changed (pr dscr) to (pr-str dscr) to fix
the ensuing NPE.

line 315: The connection is now a map with a ::web-socket key instead
of just the raw web-socket. When I made that change, I didn't update
this part. Or maybe I just forgot that I'd made that change when I
wrote this part.

core.cljs 242: Not all messages will go to a World's Worker. Refactor
that part out as preparation for adjusting the rest of the pieces
involved in this handshake to account for this.