Clean up the "World Forking" REQ/ACK a bit (WIP)
This is a 1-step-forward-2-back kind of commit.

Moved the ::action into the envelope so the browser can use that to
route messages appropriately. This led to the realization that the
envelope building really needs to be its own function.

Started splitting up the browser's message handler along those
lines. With a condp for both ::ack-forking and ::forward. I know I
need them both, but this approach really doesn't seem sustainable.

Ultimately, the browser's ::forking message is now undeliverable
because I'm trying to send it to the world's public key rather than
the currently hard-coded Session ID.