Send a ::forked signal from bottom of Worker (WIP)
...to let the Client know that it's OK to start sending messages to
update the display.

This doesn't actually reach the client. But it's forward progress.

In order to get here, I had to revamp the message exchange between the
core controller and the WorldWorker(s).

Added an extra layer to that to let them signal what kind of message
is involved (rather that just assuming that core sends DOM events and
Worker sends rendering instructions).

Converted that interchange to transit.

Renamed Worker's (main) function to (render) to reflect what it's
doing.