Debug new web server API World Fork hardler
There are a lot of brittle pieces involved here. Mostly in terms of
things like key formatting and serialization.

It still isn't fully baked: I'm not verifying the "Initiate" packet's
signature. The fact that the request is signed at all is getting a bit
ahead of myself, though.

This version renders the browser UI and then the web server correctly
fails to handle the ::forked message (because that dispatcher hasn't
been written yet).

This commit is big and hairy enough that I don't mind those defects.