Get Cookie to web server on ::forked
This was really a combination of
a) fixing last night's browser-side FSM issues
and b) associating the Cookie with the proper Worker state on the
browser so I could send it back to the web server with the ::forked
message.

This gets me back the commit from a few days ago where the web server
is ready to signal the Client to start its half of things (Q: should
any of them run on the web server side?), and I realized that I needed
the cookie then and there.