Connect Client log-propagation World
This is getting close. Everything starts up, and the Client starts
sending messages toward the web worker.

This is failing on the web side because it can't find the
corresponding World.

Still, a lot of pieces slid into place in order to make this forward
progress, and it seemed like a good time for a milestone commit.

I think this is a pretty decent API, although the way I'm using it is
over-simplified.

The system registers the Client-connect piece for the Renderer to call
at startup. Then it creates that world.

When the browser sends a ::forked message, the renderer.lib looks up
that connector function in the registrar and calls it with a function
to wire it all together (along with a parameter to use to signal that
the World is disconnecting this particular Renderer).

Converted the Client code to make it play nicely with this
approach. In particular, it sends that disconnect signal to all
connected Renderers when I halt! the system.

Moved the renderer.lib's on-message function to share
do-wrap-message's Lamport clock. At least halfway expect this to cause
issues with function declarations when I restart the JVM.