Debug previous commit (WIP)
This looks like a much bigger change than it actually is.

On the browser side, added a handler for the ::disconnect message.

On the web server side, added a couple of arities to the
message-creation functions to allow sending messages without a body.

Also added a piece to the ::forked handler on the server side to
actually send the ::ack-forked message.

This still fails in a very similar way. Now the world-key is getting
moved over to the ::active part of the world tree, but the associated
value is nil. So the errors almost look identical.

Committing here, because this looks like a big change. And I'm about
to make a change that probably will be fairly drastic. So I'm trying
to keep that one from looking even nastier than it must.