Start implementing networking pieces (WIP)
This started with a UDP socket that I can pass to both the Server and
Client Components.

Then pieces started failing in unusual ways that seemed to have
something to do with library versions that I was picking up inside the
docker container.

Cleaned up those pieces, mostly, by switching that version.

Bumped to new clojure version along the way.

Cleaned up the Dockerfile a bit.