Debug system start/stop
Whatever the problem was with using the byte-stream library got fixed
along the way.

Most of this was tweaking where definitions wind up getting passed
around among the System initialization pieces.

Had to add a secret key file for the server, because of the way that's
implemented.

Changed the startup code slightly so it doesn't try to start the
system automatically. This sped up iteration time on that drastically,
because Docker no longer needs to download everything each time a file
in the backend directory (which includes the System definition)
changes.

Incremented some dependency versions. The two that really mattered
were a) switching to the latest version of com.frereth/cp and b)
trying a newer version org.webjars/bootstrap. The latter wasn't
compatible, so rolled that experiment back until I have more time to
really look into what's changed.

This version seems to start and stop the server. But there's an
error from core.async at startup, and I don't see any logs. And it's
pretty much completely untested.

But there are enough changes here that I really wanted to commit them
before I took any further steps forward.