Get the CurveCP Server Component to start
At least the logs look like that's working now.

The finickiest part of this was setting up how/where the logger gets
set up, because I don't want to start it twice.

Specifying the UDP socket's port was also a trouble spot, because I
originally tried to do so in multiple places. So I had to track down
those false starts to find the proper place to set it.

That also led to catching exceptions when the Server startup fails. I
got tired of restarting the JVM to free up the orphaned socket owner.

Also updated the README to reflect what I'm actually doing at startup.

Added a handler to the transit serializer for Many-to-Many core.async
channels. I shouldn't be logging those, but I also don't want people
writing log messages to get hung up on what the can and cannot include
in the data portion (there's a lot of room for improvement here).