Debug Client Component startup (WIP)
Looks like a bigger change than it actually was. Rearranged the user
ns in an attempt to keep the pieces segregated.

This really boiled down to where the logger is actually stored in the
Server component.

That was trying to track down an error about missing the directory for
storing the nonce key file.

I'm really not happy with the way either of those currently work.