Debug Client startup problems (WIP)
This really amounted to adding transit serializer placeholders for
Atoms and Agents.

Now it successfully builds the error log messages about the missing
nonce key folder.

And then it fails to send them because that World connection is still
Pending. That's a state machine issue in code that really isn't
related to the current branch.