Debug disappearing log messages (WIP)
It looks as though the problem is actually that my active browser
session/connection disappears. Possibly on a refresh whenever I save
something and it automatically reloads.

Yes, it took a significant amount of work to get back to the point
where I remember enough about the way this works to draw that much of
a conclusion.