Rearrange preliminary Session State (WIP)
Initial version had this at the root of the Sessions map. This let me
set up a single ::pending Session and then copy it into ::active
without needing to worry about how it gets into ::pending in the first
place.

This combined with a world-state a couple of steps down that tree.

This was confusing enough that I started rearranging it before I
remembered why I'd set it up this way in the first place. It seemed
like it had to be wrong and part of the problem.

That part seems like it was actually working, but it's worth cutting
this knot and making the change. It has to be done sooner or later,
and I'm no longer in "rush out a demo" mode.

I've really just started by updating the specs and changing around
preliminary hard-coded bits.