Trickle out on-message! lamport param change (WIP)
This version does not build. Even so, it seemed as though there are
enough changes to justify a way-stone commit.

Working on ramifications from previous commit.

That added the parameter to the function call.

This worked those changes up the call stack until I reached the web
route creation. Converted that to a function so the web server
Component can pass it the Lamport clock Component.

Then I went back to the previous task, which was switching the way I'm
tracking state in each Session/World.

Started throwing exceptions in various places where I'm just handling
that incorrectly as reminders to fix them.

renderer.lib has a dispatch! multimethod that was using the global
sessions map in the sessions ns far too liberally. Converted that to a
parameter and started isolating where the fact that it's stored in an
atom matters (by pulling it out of the atom before I ever call
dispatch! and converting that call to happen inside swap!)