Extend Lamport Component use (WIP)
Especially in the functions that actually build the messages to send
to the browser. They were just using lexical closures with a counter
atom at the top level.

At least in theory, I think this version has most of the recent
refactoring cleaned up.

It fails immediately at startup, because I'm not creating a
Principal (or Subject...this is still up in the air) to pass along to
the log-in call to go with the fake hard-coded test session.