Start fixing renderer.lib/dispatch return vals (WIP)
Since I'm using it as the function in swap! calling it dispatch! and
ignoring the return value was totally inappropriate.

The browser connection fails when it tries to call login-finalized!
because the session atom has been replaced with a manifold.deferred.

That may just be a fluke in my current JVM life, but I don't think
I've done anything to fix it.