Finish debugging lifecycle startup
This is a fairly hefty change that touched a lot of parts, but most of
the details were obvious TODO items that I'd left strewn around as I
was setting it up.

Finished fixing up the return values for the various
renderer.lib/dispatch defmethods. I think.

The connection between the browser, web socket, web worker, and web
server seems to be back into the same shape as when I started this
refactoring 3 weeks ago.

Most of this was just finishing up dangling refactoring threads.

Moved some of the specs into a shared ns with a note that it's more
generally applicable for the renderer.

Moved some world-management functionality from renderer.lib into the
shared world ns.

Alphabetized the public functions in the sessions ns.