Update library dependencies (WIP) (broken)
And pull the rest of today's commits back together enough to get back
into a REPL.

There's a chicken/egg issue here. I've broken things badly enough
tonight that the browser just loads a seemingly empty page again.

I've also spent enough time fighting bigger-picture breakage that I'm
not sure where that fell apart.

And the purpose behind this branch *is* to try to sort the browser
code into a semblance of order, so this is a fitting start.