Debug recent changes (WIP)
Most of this was really fairly minor things like arity problems,
comment improvements, and adding debug code.

Lots of comments about deciding what/how to do different things in
terms of the World/Session FSM.

I did fix a few real issues with the way my refactored world FSM state
changes work. I added a constructor to build the preliminary state,
and fixed the parameter in the (transition) call (it was using the
original world-map rather than the updated one, so all the metadata
that got added in previous steps was discarded).

Moved the cookie out of the world's internal-state, because it belongs
to the outer metadata communications pieces.

Realized at the end that I've managed to botch the session-map atom.

Instead of a dictionary of session-keys to session instances, it's
morphed into a single session.