Start splitting ws event handlers into new ns (WIP)
I realized that I had a chicken/egg problem. The Session Manager
relies on the web socket. But the web socket event handlers really
need access to the Session Manager.

So started refactoring them into their own namespace.

This definitely is not even a complete first draft.

There are still a couple of obvious compiler errors, referencing
symbols that I haven't refactored over yet.

But there's enough here that it seemed like a good point to make a
milestone commit.