Debug the acking-fork "missing world" problem
It was a misleading error message. The actual problem stemmed from
missing the async channel used to trigger the action when that message
arrives.

This version tries to trigger the worker. Doing so fails because the
URL it uses is broken.

It fails in the wrong way because we get back a 204 rather than 404.