Try to fix issue where nil worker triggers 204 (WIP)
It's querying for http://localhost:10555/://?initiate=...

That returns a 204 rather than the 404 that would be appropriate.

Tried to adjust the routing to fix the problem. The quick/obvious
approaches based on the bidi docs didn't work.

Settle for a comment that my desire to switch to Pedestal is higher
priority than I was planning.