Start debugging the Base64 decode issue (WIP)
I was sending a keyword in the cookie that I'm trying to decode. The
keyword basically amounts to "What should this be?"

I was setting that cookie in a couple of different spots for the
world. Both during the ::forking and ::forked handlers (from the
Worker).

But I don't think the Worker has access to it anywhere.

Which probably doesn't matter, but I'm no longer certain.

Started trying to document the message flow. It's complex enough that
I probably want/need a big whiteboard to track down all the moving
parts.