Write more about the handshake
Make sure that marking a world in the forking state updates the
session-manager's world-atom rather than getting silently discarded.

Tweak the parameters around build-worker-from-exported-key so we only
need to convert the exported public key from javascript to cljs once.

Rearrange that function so we send the forking notification to the web
server before we actually trigger the fork.