Take one final visual scan through code
Found a refactoring/error-handling TODO item that needed to move from
renderer.lib to connection. That took some debugging.

Bumped library dependency versions.

Punted on the connection-test symbol naming weirdness. Sometimes it
seems to work with the leading dot, but not always. Switch it to what
I want/expect and just move on for now.