Work on debugging pedestal conversion (WIP)
Think I have the bulk of the routes translated reasonably. It's tough
to be sure: the content-type negotiator interceptor was throwing an
NPE the last time I got to it.

And my halt! function wasn't actually stopping the web server. I think
I have that fixed, but I haven't had time to test it.