Get basic requests working in pedestal (WIP)
This is still far from thoroughly tested. But the basic idea is
working again. I can call both index and echo pages.

The problem was that the content-negotiation interceptor needed me to
add the :uri key from the Ring request into the Pedestal request map
when I do that conversion.