Start digging into content-type header (WIP)
The "default" interceptors include a content-type Ring middleware
thing. It tries to determine the appropriate value based on the file
extension, which is obviously completely useless (and seems nuts) for
the root node.

Added a content-negotiation middleware which basically amounts to
crossing my fingers and hoping for the best.

I really should be paying attention to the values it sets up.

Or maybe I should just ignore that content-negotiation completely: the
index *is* going to return HTML.

I think this is forward progress, but ended by needing to restart REPL
because the compiler returned nonsense when I tried to call (reset)