Open up CSP (ugly)
Switch to CSP headers that almost allow any/everything.

This is bad.

And I'll probably have to open it even further when I get back to the
point of loading a web worker.

But this is basically what Pedestal supplies as a default. I'm going
to defer trying to secure this into its own feature branch.