Start debugging pedestal websockets (WIP)
I think I have the problem isolated. At the very least, aleph is
returning a manifold.Deferred. Pedestal sees that it isn't a
core.async channel and assumes that means that it's a synchronous
response.

Still have to prove this and try to fix it, but I think my hypothesis
is pretty sound.