Continue debugging websocket (WIP)
Biggest problem with the previous commit stemmed from what I was
returning from the websocket handler.

It was a Deferred. The secure-headers interceptor failed when it tried
to assoc onto that.

I've played with a couple of variants on there. Including adding an
interceptor that converts Deferred responses (and response bodies)
into core.async channels.

This approach, which does that immediately, seems to work better.

That's not to say that what I'm returning now, or how I'm handling it,
or not, is right.

There are longer any obvious errors on the server side.

The only hint that anything's wrong (aside from the browser not
rendering anything) is a warning message asking about a better place
to call session/do-mark-forked.

Followed by an unchecked runtime error when "The message port closed
before a response was received."