Get log-viewer loading again
Realized that the latest problem wasn't the web-socket. It was an
empty response from the /fork?initiate "handler."

The real problem was that the wrapper had gotten converted to an
interceptor, but the actual functionality was still returning a
Response rather than assoc'ing that into the Context.

The rendered view doesn't look like I remembered. It's been a while
since I've seen it, but I suspect this version may be more
correct (since it looks like it's styled more correctly).

I'll worry about debugging the visuals after I'm happy with the
functionality.