Refactor chunks of renderer.lib into handlers (WIP)
Start converting the basic idea into a data-down/events-up model.

The handlers are basically safe, with access to mostly-immutable
values that they need to be able to access. If they need to mutate the
system, they can post messages to the event bus.

Added Components to the System for both the event bus and the
subscribers.

Have at least a first draft implementation of this for the
:world/forked handler.