Find/fix disappearing session
The real problem was where I wired up the event handlers. I was
calling reduce over all the sessions instead of just the session I
want to update.

Added the event-bus to that session so I'll have it available for
debugging later.

Also, fixed a bug in event-bus/publish! where I was trying to swap!
the log atom without the atom parameter.