Tracked down/fixed disappearing ack-forking message
The problem really stemmed from a namespace mismatch on the defhandler
macro's session-atom. It threw a NPE when I tried to extract it.

This version is still broken: it throws an ArityException when I try
to call post-message! But this is a bug that's been very difficult to
track down, so I wanted to commit it immediately.