Find that disappearing session-id
The session-extractor interceptor deliberately removes it from the
Context, on the theory that anything that needs it can just find it in
the individual Session that it adds.

The problem was that I never added that in there. Moved the temporary
test-key global from sessions to connections so I can just stuff it in
there right off the bat when I initially create the session.

There was another issue with keyword namespacing mismatches for it:
sometimes I was using :frereth/session-id. Others,
::connection/session-id.

This version round-trips a valid cookie and returns the code for the
worker to render.