Start debugging the Server startup
There were a few fairly obvious problems.

Converted a bit of the logging from println to weald.

Now that I'm past the obvious errors, it just tries to start and then
the REPL quits responding with no output until I send it a SIGINT.

I'm very skeptical about whether this is a sensible part of this
branch/PR. I don't think I changed anything in this vicinity, so it's
probably a problem that already exists in master.