Clone the fulcro3 template (broken)
Based on recommendation to just start w/ fulcro3 instead of wasting
time learning the intricacies of 2.

Renamed the default apps folder to tracker and updated the ns
declarations appropriately. I think.

This version doesn't actually compile. There are problems with fulcro
3, alpha 22.

Decided that, for my present purposes, it doesn't make sense to try to
fix that.