Fix a couple of Arity exceptions
Fulcro/pathom interop.

The main problem is that fulcro's api-handler middleware expects an
arity-1 parser. The Pathom parser that I'm building is arity-2. It
requires arity-2.

The 1st parameter is called `env`, but the :muuntaja/request out of
the handler's context parameter seems to work.

Well, it works better.

The response I'm sending back to the browser (which is proper transit,
according to the MIME-type response header) fails to parse because it
isn't valid JSON.

But, hey, this is several leaps and bounds ahead of what I had an hour
or so ago.