Debug broken SESSION response (WIP)
sonic-forces/tracker

The browser and web server are not communicating correctly.

It looks to me as though the server is sending EDN while the browser
fails to read it as JSON. Even though both should be communicating in
transit.

While I debugged this, I started moving server pieces into a separate
handlers ns.