Remove Content-Type handler from raw response
For the Sonic Forces Tracker app.

This allows whatever is supposed to do the content negotiation and
serialization to actually do its job.

Now the data getting back to the browser looks correct, and the basic
login page loads without errors.

This involved copy/pasting most of Fulcro's middleware API handler so
I could leave out the part that sets that header.