Fix routing to get to the player/Root
SFT

The timing on this is wrong, since it really should load the account
information first: we don't have any idea which Player we should be
showing yet.

And the account/id is nil, so I'm still missing fundamental details
about how the routes are supposed to interact.

But it's forward progress.