Stop beating my head against this wall (WIP)
SFT

I'm hit some fundamental structural differences between the way Fulcro
3 works and the way the Fulcro 2 template was designed.

They are totally reconcilable, but it really needs a bigger-picture
step back. This particular feature branch has dragged on far too
long. My goal was to pull in my basic original functionality to try to
get some reuse out of it.

This more or less does that. Except for server interaction. And the
way the dynamic routing works. And, really, the whole authentication
thing.

But, hey, it shows a form that you can start interacting with.

The "only" problem with that part of it is a bunch of missing/nil
idents.