Debug errors about player/Root nil will-enter
SFT

Since I don't have a version defined, the default just does a
route-immediately based on my Root Component's ident. The way
everything is currently set up (at least when logged out), there is no
player ID. So it gets a nil and fails in pretty much exactly the same
way it did when I was overriding it.