Turn :player/id into a direct player/Root prop
The way this is written (which feels wrong), I still need to feed it
in from the root/Root. I think the router should pick it up
automatically.

Honestly, most of this is just me deleting TODO notes and trying to
get back into the swing of working on this project.

SFT