Try to create log-viewer mimicking system (WIP)
I want many/most of the pieces I put into place for that. Now I'm
trying to remember how they fit together, figure out how/if they make
sense in terms of shadow-cljs, and get them to compile.

Right now, it doesn't even manage that much.

This is really just an incremental check-in for the sake of my
forward-progress habit.