Split window-manager out of core (WIP)
The name isn't quite right. But, conceptually, that represents what I
want it to do.

The main point is that it is its own Component. I should be able to
manage it even if connecting the websocket fails. Because, at this
stage of the implementation, I don't have any interest in running a
server to make that connection succeed.

Spent a lot of time thinking through the details about how to render
to a texture in the WebWorker thread and then transfer that to a
texture that gets rendered in the main thread.

This may be/probably is an absolutely terrible idea/approach.

Started working through basic details behind making it work so I can
find out.

One big piece of this is the various WebWorker event handlers. Set
those up as a multi-method with a wrapper to handle Lamport clock
ticks.

Added a shared UI ns for handling things like window resizing.

The code for the main thread compiles.

The code for the Web Worker does not.

For #roadblocks

Yes, this was an extremely productive day.