Work on scene loading failures (WIP)
This is really all about the "window-manager."

One major missing piece was that I was missing the render! function.

Oops.

For whatever reason, it is not displaying animation updates from the
Worker. There are lots of possible explanations. I'm trying to track
down the correct one.

There are multiple prongs of attack going on here. I don't know a good
way to split them up.

One piece of this involved moving specs into a shared-ui ns to reduce
duplication.

Another part involved refactoring the outer "window-manager" level
scene creation functions.

Added a checkerboard floor to help visualize what's going on. Well,
attempted to do so. It isn't showing any checks that I can see, and it
disappears when I actually lay the plane down flat. So, right now,
it's sitting at about a 45 degree angle.

Another piece involved adding a send-to-worker! wrapper to my
shared/worker ns. The biggest point (for now, and aside from the basic
desire to hide implementation details) to this is to make sure I don't
forget to pass along the current Lamport tick.

for #roadblocks