Debug why web worker isn't starting
Still need to decide what to do about it.

Initial problem is that it got triggered after the authentication
handshake over the web socket.

Bypassed that to just start immediately.

This fails because it depends on having a World set up into a Pending
state sometime during that handshake.

for #roadblocks