Decide that I need a mock websocket. Etc.
I have to either a) mock it out in the browser b) add a server to
handle it or c) gut the code that uses it.

Option c) may be the correct approach. But I'm not going to undo all
that previous work until I'm sure.

That leaves option a) as the quickest way to move forward.

And also:

Tweak README instructions as I restart and actually use them.

Look into alternative approaches for coping with Firefox's lack of
Worker.requestAnimationFrame.

Lots of notes about the holes in my plans for a "Window Manager"
approach.

Along with other nips/tucks and comments as I remember how the current
approach works.

For #roadblocks