Started splitting up system initialization (WIP)
It has to happen in stages. Truly interesting stuff happens after
authentication. But there isn't any reason for the
pre-authenticated/anonymous "Attract" screen to be boring.

For a static demo site, that's where everything happens.

There's a lot going on in here. More than I'm comfortable with in a
single commit. But it's the first time in several days that I felt
like things were worth committing. As-is, this is pretty broken (I'm
back to a black canvas and an initialization error).

Adjusted the roadblocks/window-manager to use plain ctor/dtor!
functions rather than integrant. I don't want to impose Component
library choices on app authors.

Created a shared/window-manager ns with a split between ::base and
::interface, even though it smacks of OOP. They're both really just
wrappers around the "concrete" app-supplied implementation. Which also
smacks of OOP. Better approaches aren't coming to mind at the moment.

Updated the session-socket to have a reference to that shared base
window-manager. This feels like it may be a mistake.

For #roadblocks