Get basic system loading again
The fundamental problem was mostly what I noted in the previous commit
message. The arguments to the concrete window-manager ctor were
conflated. I had an ::interface with a required reference to itself.

Updated those semantics so an ::interface has an optional
::implementation. I'm still not exactly happy with this approach, but
at least it functions as well as it did before I started mucking
around with the System/Component lifecycle.

I'm back to a checkerboard floor and a spinning cube that's waiting
for the Web Worker to load.

Fixed a problem where the concrete window-manager wasn't picking up
changes during a hot reload.

for #roadblocks