Fixed GL Renderer trying to create canvas
I was passing in the off-screen one in incorrectly.

This change (which is the actual point) is really quite small, and
easy to miss among the rest of it.

Also tweaked the docs by moving the "short version" portion closer to
the top to try to make it more obvious that that's really all you
generally need.

And moved one of the handle-incoming-message! methods into
alphabetical order.